More fancy wall-laying
Only when plonking down walls: Will plonk down the correct wall corresponding to the neighbors the wall has. The neighbor walls don't update though. Also there's a glitch related to the slime's position relative to the place the wall gets built. It comes out as a north-connected wall no matter what the conditions actually are